PIV-4 improved pairing_code processing
Pairing code is optional, and only used when creating a VCI over
contetless reader. It can be provided via env PIV_PAIRING_CODE
or in opensc.conf.

In any case the paring code, if provided, must be 8 ASCII digits.

There is no not easy way to tell the user the code is invalid.
"piv_parse_pairing_code" is added to check the length and  digits.
The caller will add a debug log entry if it is invalid so
there is a record of the failure.

 On branch PIV-4-extensions
 Changes to be committed:
	modified:   src/libopensc/card-piv.c