Piv duplicate sc_atr entry removed
Remove duplicate entry in "sc_atr" for card that may support 800-73-4

With 800-73-4 Secure Messaging the SELECT AID response specifies
which cryptographic algorithms under tag 0xAC are supported for Secure Messaging.

The code was using the discovery object to test if the PIV applet is active
as some cards have a card issue of losing the login state if the SELECT AID
is used instead. (None of these cards support SM so reading the discovey object
was as good as doing SELECT_AID.)

The problme was found while running in contactless mode, card would work the first time
becaus the discovery object would not find the PIV applet so a SELECT AID was
done and it would also update the the cryptographic algorithms.

When run a second time, reading the discovery object would work but the SELECT AID
would only be done near the end of match routine for card types the may support 800-73-4.
The duplicate "sc_atr" was not listed as one the need to have SELECT AID done.

 Changes to be committed:
	modified:   card-piv.c

 interactive rebase in progress; onto 076a8856
 Last commands done (15 commands done):
    pick 651d9911 PIV-4 improved pairing_code processing
    pick 420bd554 Piv duplicate sc_atr entry removed
 Next commands to do (10 remaining commands):
    pick ca482348 PIV change processing of CVC certificates
    pick a11256e9 PIV when built without OpenSSL or older versions
 You are currently rebasing branch 'PIV-4-extensions' on '076a8856'.

 Changes to be committed:
	modified:   src/libopensc/card-piv.c